requirements.txt
langchain==0.3.23
langchain_community==0.3.21
langchain_openai==0.3.14
load_dotenv==0.1.0
numpy==2.2.5
pandas==2.2.3
python-dotenv==1.1.0
replicate==0.31.0
Requests==2.32.3
scikit_learn==1.6.1
scipy==1.15.2
SQLAlchemy==2.0.27
streamlit==1.44.1
torch==2.6.0
tqdm==4.67.1
~treamlit==1.12.0
